Use spaces, not tabs.
Tabs are an acceptable alternative to spaces in C syntax when the desired indent size is 8, as that is the official and UNIX-standard definition in spaces of tabular width.  However, schibo and Rice were hitting the tab key in Microsoft Visual Studio--an IDE which customarily offers Microsoft's simplified display of tab characters as 4 spaces wide (which is the original indent size for syntax in C-based languages, after all) even though the actual character itself is by definition to be interpreted as 8 spaces wide within any other program not overriding the definition for coding convenience purposes.  Because of this, the indentation and text alignment in the code looks off in text editors besides Microsoft Visual Studio unless all instances of the tab character are replaced with the correct number of spaces as the original authors had intended their code to be viewed using.

Either a) purposely desire 8-space indent sizes for your C code instead of 4, b) use 4-space indents for single indents and tab characters for double indents, or c) just indent each level by using 4 spaces.  Since the original authors incorrectly attempted usage of the tab character thanks to Microsoft's IDE, solution c) has been chosen here, resulting in the extinction of the tab character from all 1964 source code files in this version.