Default to using the interpreter CPU instead of the dynarec.
Unfortunately the dynarec in 0.8.5 rarely seems to work.  I think fixing it isn't as interesting as the interpreter in this version, which gets many ROMs working that don't work in the latest version of 1964 or other emulators.

This change isn't being done because of my own personal preferences or anti-dynarec campaigns so much as the fact that bringing this version of 1964 to a 64-bit and making it portable is greatly slowed down by constantly having the setting default to use a dynarec which wasn't really working from the beginning anyway.